+++ tests/testthat/test_MatchedSpectra.R
@@ -228,6 +228,13 @@ test_that("plotSpectraMirror throws an error", {
                                                         score = numeric()))
     plotSpectraMirror(ms[1])
     plotSpectraMirror(ms[1], scalePeaks = TRUE)
+
+    ## test the helper function
+    dots <- list(ppm = 15)
+    pl <- as.list(param <- CompareSpectraParam())
+    expect_equal(.res_setting(dots = dots, pl, "ppm", 20), 15)
+    expect_equal(.res_setting(list(), pl, "ppm", 20), 5)
+    expect_equal(.res_setting(list(), list(), "ppm", 20), 20)
 })
 
 test_that("addProcessing works", {