+++ NEWS.md
@@ -1,5 +1,10 @@
 # MetaboAnnotation 1.13
 
+## Changes in 1.13.2
+
+- For `plotSpectraMirror,MatchedSpectra` method, we now pass the available 
+  `ppm` and `tolerance` value that were used in the creation of the object. 
+
 ## Changes in 1.13.1
 
 - Add parameter `matchedPeaksCount` to `CompareSpectraParam` that enables